Separate building the builtins from AST, add userdata to type members
